+++ infrastructure/modules/networking/main.tf
@@ -32,11 +32,14 @@ resource "aws_subnet" "public_subnet" {
 }
 
 resource "aws_subnet" "private_subnet" {
+  for_each = var.private_subnet_cidr_block
+
   vpc_id                  = aws_vpc.this.id
-  cidr_block              = var.private_subnet_cidr_block
+  cidr_block              = each.value
+  availability_zone       = each.key
   map_public_ip_on_launch = false
 
-  tags = merge(local.tags, tomap({ "Name" = "${var.namespace}-Private-Subnet-1" }))
+  tags = merge(local.tags, tomap({ "Name" = "${var.namespace}-Private-Subnet-${each.key}" }))
 
 }
 
@@ -72,7 +75,7 @@ resource "aws_route_table_association" "this" {
 
 # Creating an Elastic IP for the NAT Gateway!
 resource "aws_eip" "this" {
-  vpc = true
+  domain = "vpc"
 
   tags = merge(local.tags, tomap({ "Name" = "${var.namespace}-eip" }))
 
@@ -106,6 +109,7 @@ resource "aws_route_table" "private_rt" {
 
 # Creating an Route Table Association of the NAT Gateway route table with the Private Subnet!
 resource "aws_route_table_association" "private_rt_association" {
-  subnet_id      = aws_subnet.private_subnet.id
+  for_each       = aws_subnet.private_subnet
+  subnet_id      = each.value.id
   route_table_id = aws_route_table.private_rt.id
 }