+++ infrastructure/modules/xc3/main.tf
@@ -51,7 +51,8 @@ resource "aws_iam_role_policy" "this" {
           "arn:aws:s3:::${aws_s3_bucket.this.id}/*",
           "arn:aws:lambda:*:*:function:*",
           "arn:aws:events:*:*:rule:*",
-          "arn:aws:events:*:*:rule/*"
+          "arn:aws:events:*:*:rule/*",
+          "arn:aws:iam::*:role/onboarding-custodian-role"
         ]
       },
       {
@@ -102,6 +103,7 @@ resource "aws_iam_instance_profile" "this" {
   tags = merge(local.tags, tomap({ "Name" = "${var.namespace}-EC2-Profile" }))
 }
 
+
 # Creating EC2 Instance that will be hosting Cloud Custodian
 
 resource "aws_instance" "this" {
@@ -109,9 +111,9 @@ resource "aws_instance" "this" {
   #ts:skip=AWS.AI.LM.HIGH.0070 We are aware of the risk and choose to skip this rule
   ami                         = data.aws_ami.ubuntu.id
   instance_type               = var.instance_type
-  associate_public_ip_address = false
+  associate_public_ip_address = var.env == "prod" ? false : true
   key_name                    = data.aws_key_pair.key_pair.key_name
-  subnet_id                   = var.subnet_id
+  subnet_id                   = var.public_subnet_ids[0]
   vpc_security_group_ids      = [var.security_group_ids.private_security_group_id]
   iam_instance_profile        = aws_iam_instance_profile.this.name
   user_data = templatefile("${path.module}/startup-script.sh.tpl", {
@@ -139,25 +141,6 @@ resource "aws_instance" "this" {
   }
 
   tags = merge(local.tags, tomap({ "Name" = "${var.namespace}-EC2" }))
-}
-
-# Creating Bastion Host Server
-
-resource "aws_instance" "bastion_host" {
-  #ts:skip=AC-AWS-NS-IN-M-1172 We are aware of the risk and choose to skip this rule
-  #ts:skip=AWS.AI.LM.HIGH.0070 We are aware of the risk and choose to skip this rule
-  ami                         = data.aws_ami.ubuntu.id
-  instance_type               = var.instance_type
-  associate_public_ip_address = true
-  key_name                    = data.aws_key_pair.key_pair.key_name
-  subnet_id                   = var.public_subnet_ids[0]
-  vpc_security_group_ids      = [var.security_group_ids.public_security_group_id]
-  iam_instance_profile        = aws_iam_instance_profile.this.name
-  root_block_device {
-    volume_size = 30
-    encrypted   = true
-  }
-  tags = merge(local.tags, tomap({ "Name" = "${var.namespace}-Bastion-Host-Server" }))
 }
 
 # Configuring SES Identity and SQS for email notifications
@@ -192,16 +175,26 @@ resource "aws_s3_bucket" "this" {
 
 # Uploading Cloud Custodian Policies and lambda layers in S3 bucket
 # tflint-ignore: terraform_required_providers
-resource "null_resource" "upload_files_on_s3" {
-  triggers = {
-    s3_bucket = aws_s3_bucket.this.arn
-  }
+resource "terraform_data" "upload_files_on_s3" {
+  triggers_replace = [aws_s3_bucket.this.arn]
 
   provisioner "local-exec" {
     command = <<EOT
       aws s3 cp python.zip s3://${aws_s3_bucket.this.id}/lambda_layers/
       aws s3 cp ../custom_dashboard/grafana_dashboards/. s3://${aws_s3_bucket.this.id}/content/ --recursive --exclude "*.md"
+      aws s3 cp ../cloud_custodian_policies/ s3://${aws_s3_bucket.this.id}/cloud_custodian_policies/ --recursive --exclude "*.md" --include "*"
    EOT
+  }
+}
+
+# tflint-ignore: terraform_required_providers
+resource "terraform_data" "eicendpoint" {
+  count = var.env == "prod" ? 1 : 0
+  triggers_replace = [
+    aws_instance.this.id
+  ]
+  provisioner "local-exec" {
+    command = "aws ec2 create-instance-connect-endpoint --subnet-id ${var.private_subnet_id[0]} --security-group-id ${var.security_group_ids.private_security_group_id}"
   }
 }
 
@@ -214,6 +207,6 @@ resource "aws_lambda_layer_version" "lambda_layer_prometheus" {
 
   compatible_runtimes = ["python3.9"]
   depends_on = [
-    null_resource.upload_files_on_s3
+    terraform_data.upload_files_on_s3
   ]
 }