+++ infrastructure/variables.tf
@@ -33,7 +33,7 @@ variable "public_subnet_cidr_block" {
 }
 
 variable "private_subnet_cidr_block" {
-  type        = string
+  type        = map(string)
   description = "The CIDR Block of the private subnet"
 }
 
@@ -141,4 +141,10 @@ variable "slack_channel_url" {
 variable "create_kms" {
   description = "Fetch the KMS if exist"
   type        = bool
+}
+
+variable "env" {
+  description = "Env variable for Dev/Prod"
+  type        = string
+  default     = "dev"
 }